moved validate function to LSM
