+++ x/stakeibc/keeper/lsm_test.go
@@ -175,45 +175,6 @@ func (s *KeeperTestSuite) TestShouldCheckIfValidatorWasSlashed() {
 	}
 }
 
-func (s *KeeperTestSuite) TestRefundLSMToken() {
-	liquidStakerAddress := s.TestAccs[0]
-	despositAddress := types.NewHostZoneDepositAddress(HostChainId)
-
-	// Fund the module account with the LSM token
-	lsmTokenIBCDenom := "ibc/cosmosvalXXX"
-	stakeAmount := sdk.NewInt(1000)
-	lsmToken := sdk.NewCoin(lsmTokenIBCDenom, stakeAmount)
-	s.FundAccount(despositAddress, lsmToken)
-
-	// Setup the liquid stake object that provides the context for the refund
-	liquidStake := types.LSMLiquidStake{
-		Staker: liquidStakerAddress,
-		HostZone: types.HostZone{
-			DepositAddress: despositAddress.String(),
-		},
-		LSMIBCToken: lsmToken,
-	}
-
-	// Refund the token and check that it has been successfully transferred
-	err := s.App.StakeibcKeeper.RefundLSMToken(s.Ctx, liquidStake)
-	s.Require().NoError(err, "no error expected when refunding LSM token")
-
-	stakerBalance := s.App.BankKeeper.GetBalance(s.Ctx, liquidStakerAddress, lsmTokenIBCDenom)
-	s.CompareCoins(lsmToken, stakerBalance, "staker should have received their LSM token back")
-
-	moduleBalance := s.App.BankKeeper.GetBalance(s.Ctx, despositAddress, lsmTokenIBCDenom)
-	s.True(moduleBalance.IsZero(), "module account should no longer have the LSM token")
-
-	// Attempt to refund again, it should fail from an insufficient balance
-	err = s.App.StakeibcKeeper.RefundLSMToken(s.Ctx, liquidStake)
-	s.Require().ErrorContains(err, "insufficient funds")
-
-	// Attempt to refund with an invalid host zone address, it should fail
-	liquidStake.HostZone.DepositAddress = ""
-	err = s.App.StakeibcKeeper.RefundLSMToken(s.Ctx, liquidStake)
-	s.Require().ErrorContains(err, "host zone address is invalid")
-}
-
 func (s *KeeperTestSuite) TestDetokenizeLSMDeposit() {
 	// Create the delegation ICA
 	owner := types.FormatICAAccountOwner(HostChainId, types.ICAAccountType_DELEGATION)
@@ -239,6 +200,7 @@ func (s *KeeperTestSuite) TestDetokenizeLSMDeposit() {
 		Denom:   denom,
 		Amount:  sdk.NewInt(1000),
 		Status:  recordstypes.LSMTokenDeposit_DETOKENIZATION_QUEUE,
+		StToken: sdk.NewCoin(StAtom, sdk.OneInt()),
 	}
 	s.App.RecordsKeeper.SetLSMTokenDeposit(s.Ctx, initalDeposit)
